4688 removing varMap v1
