+++ src/analysis/analysisVisitors/varAnalysis/varMap.test.ts
@@ -15,332 +15,160 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import VarMap, {
-  mergeExistenceMaps,
-  SELF_EXISTENCE,
-  VarExistence,
-} from "./varMap";
+import VarMap, { VarExistence } from "./varMap";
 
-describe("VarMap", () => {
-  let varMap: VarMap;
-  beforeEach(() => {
-    varMap = new VarMap();
+describe("setting output key", () => {
+  // Use case: setting the existence for a block's output
+  test("sets the existence", () => {
+    const varMap = new VarMap();
+
+    varMap.setOutputKeyExistence(
+      "brick1",
+      "@foo",
+      VarExistence.DEFINITELY,
+      false
+    );
+    varMap.setOutputKeyExistence(
+      "brick2",
+      "@bar",
+      VarExistence.DEFINITELY,
+      false
+    );
+    expect(varMap.isVariableDefined("@foo")).toBeTrue();
+    expect(varMap.isVariableDefined("@bar")).toBeTrue();
   });
 
-  test.each(["foo.bar", "foo", 'foo["bar baz"].qux[0]'])(
-    "get the exact known var (%s)",
-    (varName) => {
-      varMap.setExistence(varName, VarExistence.DEFINITELY);
-      const actual = varMap.getExistence(varName);
+  // No real use case, just a functionality expectations
+  test("overwrites any previous for the same source", () => {
+    const varMap = new VarMap();
+    varMap.setOutputKeyExistence(
+      "brick1",
+      "@foo",
+      VarExistence.DEFINITELY,
+      false
+    );
 
-      expect(actual).toBe(VarExistence.DEFINITELY);
-    }
-  );
+    varMap.setOutputKeyExistence(
+      "brick1",
+      "@bar",
+      VarExistence.DEFINITELY,
+      false
+    );
 
-  test.each([VarExistence.MAYBE, VarExistence.DEFINITELY])(
-    "doesn't overwrite nested vars (%s)",
-    (existence) => {
-      varMap.setExistence("foo.bar", existence);
-      varMap.setExistence("foo.baz", existence);
-
-      expect(varMap.getExistence("foo.bar")).toBe(existence);
-      expect(varMap.getExistence("foo.baz")).toBe(existence);
-    }
-  );
-
-  test.each([VarExistence.MAYBE, VarExistence.DEFINITELY])(
-    "gets nested existence (%s)",
-    (existence) => {
-      varMap.setExistence("foo.bar", existence);
-      expect(varMap.getExistence("foo.bar")).toBe(existence);
-      expect(varMap.getExistence("foo.baz")).toBeUndefined();
-    }
-  );
-
-  test.each([VarExistence.MAYBE, VarExistence.DEFINITELY])(
-    "gets nested existence 2 (%s)",
-    (existence) => {
-      varMap.setExistence('foo["bar baz"].qux', existence);
-      expect(varMap.getExistence('foo["bar baz"]')).toBe(existence);
-      expect(varMap.getExistence("foo")).toBe(existence);
-    }
-  );
-
-  test("get a child property of a container with any schema", () => {
-    varMap.setExistence("foo", VarExistence.MAYBE, true);
-    const actual = varMap.getExistence("foo.bar");
-
-    expect(actual).toBe(VarExistence.MAYBE);
+    expect(varMap.isVariableDefined("@foo")).toBeFalse();
+    expect(varMap.isVariableDefined("@bar")).toBeTrue();
   });
 
-  test("set existence from context obj", () => {
-    varMap.setExistenceFromObj({
+  test.each([
+    [true, true],
+    [false, false],
+  ])("works when allow any child = %s", (allowAnyChild, expectedExistence) => {
+    const varMap = new VarMap();
+    varMap.setOutputKeyExistence(
+      "brick1",
+      "@foo",
+      VarExistence.DEFINITELY,
+      allowAnyChild
+    );
+
+    expect(varMap.isVariableDefined("@foo.bar")).toBe(expectedExistence);
+  });
+});
+
+describe("setExistenceFromValues", () => {
+  // Use case: setting the @input based on the page reader schema or adding existence from trace
+  test("sets the existence for a plain object", () => {
+    const values = {
+      "@foo": "bar",
+    };
+
+    const varMap = new VarMap();
+    varMap.setExistenceFromValues("brick1", values);
+
+    expect(varMap.isVariableDefined("@foo")).toBeTrue();
+    expect(varMap.isVariableDefined("@foo.bar")).toBeFalse();
+  });
+
+  test("sets the existence for a complex object", () => {
+    const values = {
       foo: {
         bar: "baz",
       },
       qux: "quux",
-    });
+    };
 
-    expect(varMap.getExistence("foo")).toBe(VarExistence.DEFINITELY);
-    expect(varMap.getExistence("foo.bar")).toBe(VarExistence.DEFINITELY);
-    expect(varMap.getExistence("baz")).toBeUndefined();
-    expect(varMap.getExistence("qux")).toBe(VarExistence.DEFINITELY);
-    expect(varMap.getExistence("quux")).toBeUndefined();
+    const varMap = new VarMap();
+    varMap.setExistenceFromValues("brick1", values);
+
+    expect(varMap.isVariableDefined("foo")).toBeTrue();
+    expect(varMap.isVariableDefined("foo.bar")).toBeTrue();
+    expect(varMap.isVariableDefined("foo.bar.baz")).toBeFalse();
+    expect(varMap.isVariableDefined("qux")).toBeTrue();
+    expect(varMap.isVariableDefined("quux")).toBeFalse();
   });
 
   test("set existence from context obj with parent specified", () => {
-    varMap.setExistence("foo", VarExistence.DEFINITELY);
-    varMap.setExistenceFromObj(
+    const varMap = new VarMap();
+    varMap.setExistenceFromValues(
+      "brick1",
       {
         qux: "quux",
       },
       "bar.baz"
     );
 
-    expect(varMap.getExistence("foo")).toBe(VarExistence.DEFINITELY);
-    expect(varMap.getExistence("bar.baz.qux")).toBe(VarExistence.DEFINITELY);
+    expect(varMap.isVariableDefined("bar.baz.qux")).toBeTrue();
   });
+});
 
+describe("cloning", () => {
   test("clones a var map", () => {
-    varMap.setExistence("foo", VarExistence.DEFINITELY);
+    const varMap = new VarMap();
+    varMap.setOutputKeyExistence(
+      "brick1",
+      "@foo",
+      VarExistence.DEFINITELY,
+      false
+    );
 
     const clone = varMap.clone();
-    clone.setExistence("bar", VarExistence.DEFINITELY);
+    clone.setOutputKeyExistence(
+      "brick2",
+      "@bar",
+      VarExistence.DEFINITELY,
+      false
+    );
 
-    expect(varMap.getExistence("foo")).toBe(VarExistence.DEFINITELY);
-    expect(varMap.getExistence("bar")).toBeUndefined();
+    expect(varMap.isVariableDefined("@foo")).toBeTrue();
+    expect(varMap.isVariableDefined("@bar")).toBeFalse();
 
-    expect(clone.getExistence("foo")).toBe(VarExistence.DEFINITELY);
-    expect(clone.getExistence("bar")).toBe(VarExistence.DEFINITELY);
+    expect(clone.isVariableDefined("@foo")).toBeTrue();
+    expect(clone.isVariableDefined("@bar")).toBeTrue();
   });
+});
 
-  test.each([VarExistence.MAYBE, VarExistence.DEFINITELY])(
-    "merges 2 var maps (%s)",
-    (existence) => {
-      const varMap1 = new VarMap();
-      varMap1.setExistence("foo.bar", existence);
-      varMap1.setExistence("qux", existence);
-
-      const varMap2 = new VarMap();
-      varMap2.setExistence("foo.baz", existence);
-      varMap2.setExistence("quux", existence);
-
-      const merged = varMap1.merge(varMap2);
-
-      expect(merged.getExistence("foo.bar")).toBe(existence);
-      expect(merged.getExistence("foo.baz")).toBe(existence);
-      expect(merged.getExistence("qux")).toBe(existence);
-      expect(merged.getExistence("quux")).toBe(existence);
-    }
-  );
-
-  test("merged map is a separate object", () => {
+describe("addSourceMap", () => {
+  // Use case: adding an output of previous brick to the current brick's available vars
+  test("adds a source map", () => {
     const varMap1 = new VarMap();
-    varMap1.setExistence("foo.bar", VarExistence.DEFINITELY);
-
-    const merged = varMap1.merge(new VarMap());
-    merged.setExistence("foo.baz", VarExistence.DEFINITELY);
-
-    expect(varMap1.getExistence("foo.bar")).toBe(VarExistence.DEFINITELY);
-    expect(varMap1.getExistence("foo.baz")).toBeUndefined();
-
-    expect(merged.getExistence("foo.bar")).toBe(VarExistence.DEFINITELY);
-    expect(merged.getExistence("foo.baz")).toBe(VarExistence.DEFINITELY);
-  });
-
-  test("merges meta source", () => {
-    const varMap = new VarMap();
-    varMap.setExistence("foo", VarExistence.DEFINITELY);
-    varMap.setSource("foo", "bar");
-
-    const merged = new VarMap().merge(varMap);
-
-    expect(merged.getMeta("foo").source).toBe("bar");
-  });
-
-  test("merges allowAnyChild flag", () => {
-    const varMap = new VarMap();
-    varMap.setExistence("foo", VarExistence.DEFINITELY, true);
-
-    const merged = new VarMap().merge(varMap);
-
-    expect(merged.getExistence("foo.bar.baz.qux")).toBe(VarExistence.MAYBE);
-  });
-
-  test("doesn't override a var with nested keys", () => {
-    // Set nested keys
-    varMap.setExistence("foo.bar", VarExistence.DEFINITELY);
-
-    // Try to override the parent key
-    varMap.setExistence("foo", VarExistence.MAYBE);
-
-    expect(varMap.getExistence("foo.bar")).toBe(VarExistence.DEFINITELY);
-  });
-
-  test("doesn't override a DEFINITELY var with MAYBE", () => {
-    varMap.setExistence("foo", VarExistence.DEFINITELY);
-    varMap.setExistence("foo", VarExistence.MAYBE);
-
-    expect(varMap.getExistence("foo")).toBe(VarExistence.DEFINITELY);
-  });
-});
-
-describe("VarMap meta", () => {
-  test("sets source to undefined var", () => {
-    const varMap = new VarMap();
-    varMap.setSource("@foo", "bar");
-
-    expect(varMap.getMeta("@foo").source).toBe("bar");
-    expect(varMap.getExistence("@foo")).toBeUndefined();
-  });
-
-  test.each([VarExistence.MAYBE, VarExistence.DEFINITELY])(
-    "sets source to an existing var (%s)",
-    (existence) => {
-      const varMap = new VarMap();
-      varMap.setExistence("@foo.bar", existence);
-      varMap.setSource("@foo.bar", "baz");
-
-      expect(varMap.getMeta("@foo.bar").source).toBe("baz");
-      expect(varMap.getExistence("@foo.bar")).toBe(existence);
-    }
-  );
-
-  test("clones source", () => {
-    const varMap = new VarMap();
-    varMap.setSource("@foo", "bar");
-
-    const clone = varMap.clone();
-
-    expect(clone.getMeta("@foo").source).toBe("bar");
-  });
-
-  test("merges source", () => {
-    const varMap = new VarMap();
-    varMap.setSource("@foo", "bar");
-
-    const merged = new VarMap().merge(varMap);
-
-    expect(merged.getMeta("@foo").source).toBe("bar");
-  });
-});
-
-describe("mergeExistenceMaps", () => {
-  test("works with nulls", () => {
-    const actual = mergeExistenceMaps(null, null);
-    expect(actual).toEqual({});
-  });
-
-  test("merges plain maps", () => {
-    const actual = mergeExistenceMaps(
-      {
-        foo: {
-          [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        },
-      },
-      {
-        bar: {
-          [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        },
-      }
+    varMap1.setOutputKeyExistence(
+      "brick1",
+      "@foo",
+      VarExistence.DEFINITELY,
+      false
     );
-    expect(actual).toEqual({
-      foo: {
-        [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-      },
-      bar: {
-        [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-      },
-    });
-  });
 
-  test("merges nested maps", () => {
-    const actual = mergeExistenceMaps(
-      {
-        foo: {
-          bar: {
-            [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-          },
-        },
-      },
-      {
-        foo: {
-          baz: {
-            [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-          },
-        },
-      }
+    const varMap2 = new VarMap();
+    varMap2.setOutputKeyExistence(
+      "brick2",
+      "@bar",
+      VarExistence.DEFINITELY,
+      false
     );
-    expect(actual).toEqual({
-      foo: {
-        bar: {
-          [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        },
-        baz: {
-          [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        },
-      },
-    });
-  });
 
-  test("merges Existence with an object", () => {
-    const mapA = {
-      foo: {
-        bar: {
-          [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        },
-      },
-    };
-    const mapB = {
-      foo: {
-        [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-      },
-    };
+    varMap1.addSourceMap(varMap2);
 
-    const expected = {
-      foo: {
-        [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        bar: {
-          [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        },
-      },
-    };
-
-    const actualAB = mergeExistenceMaps(mapA, mapB);
-    expect(actualAB).toEqual(expected);
-
-    const actualBA = mergeExistenceMaps(mapB, mapA);
-    expect(actualBA).toEqual(expected);
-  });
-
-  test("overrides the existence with more strict one regardless of the order", () => {
-    const mapA = {
-      foo: {
-        bar: {
-          [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        },
-      },
-    };
-    const mapB = {
-      foo: {
-        bar: {
-          [SELF_EXISTENCE]: VarExistence.MAYBE,
-        },
-      },
-    };
-
-    const expected = {
-      foo: {
-        bar: {
-          [SELF_EXISTENCE]: VarExistence.DEFINITELY,
-        },
-      },
-    };
-
-    const actualAB = mergeExistenceMaps(mapA, mapB);
-    expect(actualAB).toEqual(expected);
-
-    const actualBA = mergeExistenceMaps(mapB, mapA);
-    expect(actualBA).toEqual(expected);
+    expect(varMap1.isVariableDefined("@foo")).toBeTrue();
+    expect(varMap1.isVariableDefined("@bar")).toBeTrue();
   });
 });
